requirements.txt
Django==1.6.2
Pillow==2.3.0
Scrapy==0.22.2
South==0.8.4
Twisted==13.2.0
cffi==0.8.2
cryptography==0.3
cssselect==0.9.1
django-appconf==0.6
django-bootstrap-form==3.1
django-filebrowser==3.5.4
django-grappelli==2.5.2
django-jsonfield==0.9.12
django-localflavor==1.0
django-solo==1.0.4
django-user-accounts==1.0c4
django-wysiwyg-redactor==0.3.9
eventlog==0.7.0
lxml==3.3.4
metron==1.3.1
pinax-theme-bootstrap==5.1.0
pyOpenSSL==0.14
pycparser==2.10
pytz==2014.2
queuelib==1.1.1
six==1.6.1
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.1.1
